pasing hapus data

+++ tambahDataWarga/tambahData.php
@@ -129,6 +129,7 @@
                         <th>Tempat, Tanggal Lahir</th>
                         <th>Pekerjaan</th>
                         <th>Jenis Kelamin</th>
+                        <th>Status</th>
                         <th>Action</th>
                       </tr>
                       <?php 
@@ -154,9 +155,10 @@
                               <td><?php echo $data['ttl']; ?></td>
                               <td><?php echo $data['pekerjaan']; ?></td>
                               <td><?php echo $data['jenisKelamin']; ?></td>
+                              <td><?php echo $data['status']; ?></td>
                               <td>
                                 <a class="btn btn-sm btn-primary" href="form_editData.php?id_warga=<?php echo $data['id_warga']; ?>">Edit</a> 
-                                <a class="btn btn-sm btn-primary" href="deleteData.php?id_warga=<?php echo $data['id_warga']; ?>">Hapus</a> 
+                                <a class="btn btn-sm btn-primary" href="deleteData.php?id_warga=<?php echo $data['id_warga']; ?>" onclick="return confirm()">Hapus</a> 
                                 <a class="btn btn-sm btn-primary" href="detail.php?id_warga=<?php echo $data['id_warga']; ?>">Detail</a>
                               </td>
                             </tr>
